Migrate tests to individual files
I mananged to get 6/8 of all tests going green. Some small things need to be adjusted.

I also commented out console logs int he reducers that were running while testing on "createReducers"
