+++ test/test.todo.md
@@ -1,0 +1,13 @@
+ What do I want to test/make sure works with absolution
+ - redux ACTIONS are doing what they intend to do (both sides, firing and
+  components reception)
+    - [ ] UI_OPEN
+    - [ ] UI_CLOSE
+    - [ ] UI_STYLE_MESSENGER
+    - [ ] UI_STYLE_MESSENGER
+    - [ ] CHAT_CONNECTED
+    - [ ] CHAT_DISCONNECTED
+    - [ ] CHAT_MESSAGE_OPERATOR
+    - [ ] CHAT_MESSAGE_CLIENT
+ - styling is available in state
+ - ?
